Added the possibility to add a default value to the DirectoryInputDefinition model, as well as a new 'dynamic_default' field which is meant to be used to add a default that is not the interface's default and that may be formatted using the particular run's parameters. Also, changed the Output instances creation at the end of a run to skip non-existing key in the returned dictionary (instead of raising an ObjectDoesNotExist error).

+++ django_analyses/models/run.py
@@ -2,6 +2,7 @@ import shutil
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
 from django_analyses.models.input.definitions.string_input_definition import (
     StringInputDefinition,
@@ -34,7 +35,8 @@ class Run(TimeStampedModel):
         return f"#{self.id} {self.analysis_version} run from {self.created}"
 
     def delete(self, using=None, keep_parents=False):
-        shutil.rmtree(self.path)
+        if self.path:
+            shutil.rmtree(self.path)
         return super().delete(using=using, keep_parents=keep_parents)
 
     def create_input_instance(self, key: str, value) -> Input:
@@ -72,6 +74,22 @@ class Run(TimeStampedModel):
             )
         ]
 
+    def get_missing_dynamic_defaults(self, **kwargs) -> list:
+        return [
+            definition
+            for definition in self.analysis_version.input_definitions
+            if getattr(definition, "dynamic_default", False)
+            and definition.key not in kwargs
+        ]
+
+    def create_missing_dynamic_defaults(self, **kwargs) -> list:
+        return [
+            dynamic_default_definition.create_input_instance(run=self)
+            for dynamic_default_definition in self.get_missing_dynamic_defaults(
+                **kwargs
+            )
+        ]
+
     def create_missing_output_directory_definition(self, **kwargs) -> DirectoryInput:
         definition = self.get_missing_output_directory_definition()
         if definition:
@@ -82,6 +100,7 @@ class Run(TimeStampedModel):
             self.create_input_instance(key, value) for key, value in kwargs.items()
         ]
         input_instances += self.create_missing_output_path_definitions(**kwargs)
+        input_instances += self.create_missing_dynamic_defaults(**kwargs)
         output_directory = self.create_missing_output_directory_definition()
         if output_directory:
             input_instances += [output_directory]
@@ -89,24 +108,21 @@ class Run(TimeStampedModel):
 
     def create_output_path_destinations(self, inputs: models.QuerySet):
         for inpt in inputs:
-            is_output_path = getattr(inpt.definition, "is_output_path", False)
-            is_output_directory = getattr(inpt.definition, "is_output_directory", False)
-            if is_output_path:
-                path = Path(inpt.value).parent
-            elif is_output_directory:
-                path = Path(inpt.value)
-            else:
-                continue
-            path.mkdir(parents=True, exist_ok=True)
+            inpt.create_required_destination()
 
     def create_output_instance(self, key: str, value) -> Output:
-        output_definition = self.analysis_version.output_definitions.get(key=key)
-        return output_definition.create_output_instance(value=value, run=self)
+        try:
+            output_definition = self.analysis_version.output_definitions.get(key=key)
+        except ObjectDoesNotExist:
+            pass
+        else:
+            return output_definition.create_output_instance(value=value, run=self)
 
     def create_output_instances(self, **results) -> list:
-        return [
+        output_instances = [
             self.create_output_instance(key, value) for key, value in results.items()
         ]
+        return [output for output in output_instances if output is not None]
 
     def get_input_set(self) -> models.QuerySet:
         return self.base_input_set.select_subclasses()
@@ -133,7 +149,8 @@ class Run(TimeStampedModel):
 
     @property
     def path(self) -> Path:
-        return Path(settings.ANALYSIS_BASE_PATH) / str(self.id)
+        path = Path(settings.ANALYSIS_BASE_PATH) / str(self.id)
+        return path if path.is_dir() else None
 
     @property
     def input_defaults(self) -> dict:
